Fix Slack authentication in blocking client
The Slack API has moved on, while this crate has not. Adapted
the approach from https://github.com/slack-rs/slack-rs-api/issues/107
to pluck the auth token out of the parameter body and send it as
a bearer auth header instead. It is rather surprising to realize that
Slack ever accepted this as a body param, but hey.